+++ jwst/associations/tests/test_level3_spectrographic.py
@@ -51,21 +51,21 @@ class TestLevel3Spec(BasePoolRule):
             'spec3',
             'jw99009-o001_spec3_\d{3}_asn',
             'jw99009-o001_{source_id}_nirspec_f100lp-g140m-s200a2-s200a2',
-            set(('science', 'target_acquistion', 'autowave'))
+            set(('science', 'target_acquisition', 'autowave'))
         ),
         (
             'o002',
             'spec3',
             'jw99009-o002_spec3_\d{3}_asn',
             'jw99009-o002_{source_id}_nirspec_f100lp-g140h',
-            set(('science', 'target_acquistion', 'autoflat', 'autowave'))
+            set(('science', 'target_acquisition', 'autoflat', 'autowave'))
         ),
         (
             'o003',
             'spec3',
             'jw99009-o003_spec3_\d{3}_asn',
             'jw99009-o003_t002_nirspec',
-            set(('science', 'target_acquistion', 'autowave'))
+            set(('science', 'target_acquisition', 'autowave'))
         ),
     ]
 )